+++ freamon/model_selection/__init__.py
@@ -12,9 +12,12 @@ from freamon.model_selection.cross_validation import (
     cross_validate,
     time_series_cross_validate,
     walk_forward_validation,
+    create_time_series_cv,
+    create_stratified_cv,
+    create_kfold_cv,
 )
 
-from freamon.model_selection.cv_trainer import CrossValidatedTrainer
+from freamon.model_selection.cv_trainer import CrossValidationTrainer
 from freamon.model_selection.cv_training_step import CrossValidationTrainingStep
 
 __all__ = [
@@ -24,6 +27,9 @@ __all__ = [
     "cross_validate",
     "time_series_cross_validate",
     "walk_forward_validation",
-    "CrossValidatedTrainer",
+    "create_time_series_cv",
+    "create_stratified_cv",
+    "create_kfold_cv",
+    "CrossValidationTrainer",
     "CrossValidationTrainingStep",
 ]